(PUP-10606) Separate puppet/ffi from puppet/util
This commit attempts to make the `puppet/ffi/windows` tree usable
without having to load the entire `puppet/util/windows` module.

Move the APITypes module from `util` to `ffi`, since all the code there
is FFI-related.

Do a nested initialization of the `Puppet::Util::Windows::String` module
so it can be used without loading the base `Puppet::Util::Windows`
module.
