+++ lib/puppet/ffi.rb
@@ -1,2 +1,6 @@
-module Puppet::FFI
+require 'ffi'
+
+module Puppet
+  module FFI 
+  end
 end
