+++ lib/puppet/ffi/windows.rb
@@ -1,6 +1,8 @@
 require 'puppet/ffi'
+
 module Puppet::FFI::Windows
+  require 'puppet/ffi/windows/api_types'
+  require 'puppet/ffi/windows/constants'
   require 'puppet/ffi/windows/structs'
   require 'puppet/ffi/windows/functions'
-  require 'puppet/ffi/windows/constants'
 end
